# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/department/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of Deaprtment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Get Departments",
                "parameters": [
                    {
                        "type": "string",
                        "name": "depType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "department",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "departmenttypeid",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "expand",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "parent",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "parentid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response object",
                        "schema": {
                            "$ref": "#/definitions/responses.DepartmentsResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a Department (requires JWT)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Create Department",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "departmenttypeid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "parentid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 100,
                        "minLength": 3,
                        "type": "string",
                        "name": "title",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response object",
                        "schema": {
                            "$ref": "#/definitions/responses.DepartmentResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/department/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a Deaprtment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Get Department",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "name": "expand",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response object",
                        "schema": {
                            "$ref": "#/definitions/responses.DepartmentResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a Department (requires JWT)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Update Department",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "departmenttypeid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "parentid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 100,
                        "minLength": 3,
                        "type": "string",
                        "name": "title",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response object",
                        "schema": {
                            "$ref": "#/definitions/responses.DepartmentResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns No content (requires JWT)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Delete Department",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Response object",
                        "schema": {
                            "$ref": "#/definitions/resedist_internal_modules_department_department_responses.NoContentResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/sso/auth/login": {
            "post": {
                "description": "Authenticate user and return a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SSO"
                ],
                "summary": "Login and get JWT token",
                "parameters": [
                    {
                        "maxLength": 100,
                        "minLength": 3,
                        "type": "string",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 100,
                        "minLength": 8,
                        "type": "string",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/sso/auth/refresh": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Refresh an access token using a refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SSO"
                ],
                "summary": "Refresh access token",
                "parameters": [
                    {
                        "type": "string",
                        "name": "client_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "grant_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "refresh_token",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "pagination.PagePack": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
        "resedist_internal_modules_department_department_responses.NoContentResponse": {
            "type": "object",
            "properties": {
                "_error_code": {
                    "type": "string",
                    "example": ""
                },
                "_message": {
                    "type": "string",
                    "example": "null"
                },
                "_status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "responses.Department": {
            "type": "object",
            "properties": {
                "departmentType": {},
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "parent": {},
                "title": {
                    "type": "string"
                }
            }
        },
        "responses.DepartmentResponse": {
            "type": "object",
            "properties": {
                "_error_code": {
                    "type": "string",
                    "example": ""
                },
                "_message": {
                    "type": "string",
                    "example": "null"
                },
                "_status": {
                    "type": "string",
                    "example": "success"
                },
                "data": {
                    "$ref": "#/definitions/responses.Department"
                }
            }
        },
        "responses.DepartmentsResponse": {
            "type": "object",
            "properties": {
                "_error_code": {
                    "type": "string",
                    "example": ""
                },
                "_message": {
                    "type": "string",
                    "example": "null"
                },
                "_status": {
                    "type": "string",
                    "example": "success"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.Department"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/pagination.PagePack"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}